Introduce more lifetimes
So, before we had `'a` for a whole bunch of stuff, esp. in things like
`&'a Speller<'a>`. That lead to some weird borrowck errors. While I'm
sure there are more clever solutions to this, I just split those
lifetimes into two: One to indicate the scope of the reference, and one
to show the actual life time of the data.

In this relation, the data always needs to outlive the references to it,
so: `'a: 'b`.